requirements.txt
asgiref==3.4.1
Django==3.2.5
django-allauth==0.41.0
django-querysetsequence==0.14
oauthlib==3.1.1
Pillow==8.3.1
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
